+++ llvm/lib/SYCLLowerIR/CMakeLists.txt
@@ -53,7 +53,6 @@ add_llvm_component_library(LLVMSYCLLowerIR
   LowerESIMDVLoadVStore.cpp
   LowerESIMDVecArg.cpp
   LowerWGLocalMemory.cpp
-  ESIMDVerifier.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${LLVM_MAIN_INCLUDE_DIR}/llvm/SYCLLowerIR
